requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-extensions==3.2.3
pydot==1.4.2
pyparsing==3.1.1
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

## Following are the standard dependencies that i'll use for most of the servers :) if that takes too much time then avoid using them.

absl-py==2.1.0
aiohttp==3.9.3
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
awscli==1.32.48
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
boto3==1.34.48
botocore==1.34.48
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
colour==0.1.5
comm==0.2.2
conan==2.0.13
contourpy==1.2.0
cryptography==42.0.4
cycler==0.10.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.0
distlib==0.3.7
dnspython==2.7.0
docker==6.1.3
docutils==0.16
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.2
fasteners==0.19
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
flutter==0.1
flux==1.3.5
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.11.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.24.6
idna==2.10
importlib_metadata==8.4.0
inference-cli==0.9.13
inference-sdk==0.9.13
inference-tools==0.13.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.2.1
kiwisolver==1.4.5
labelImg==1.8.6
langchain==0.0.325
langsmith==0.0.92
lazycodr==0.2.3
libclang==18.1.1
lxml==5.2.1
manimlib==0.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.6.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==0.28.1
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.4
patch-ng==1.17.4
pathspec==0.11.2
pendulum==3.0.0
pillow==10.2.0
pipenv==2023.10.24
platformdirs==3.11.0
prettytable==3.10.0
progressbar==2.5
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.8
psycopg2==2.9.9
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1_modules==0.4.1
# pycairo @ file:///F:/generate_life_calendar-master/pycairo-1.26.0-cp312-cp312-win_amd64.whl#sha256=8616408ae93de4824a3777ec532ea75643e4bf74e49d601062c0b1788180c962
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
pyfiglet==1.0.2
pygame==2.5.2
PyGithub==2.3.0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.19
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.0
safetensors==0.4.4
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.16.0
sentry-sdk==1.40.6
setuptools==69.5.1
shellingham==1.5.4
sip==6.8.3
six==1.16.0
skypilot==0.4.1
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.40.0
streamlit==1.38.0
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tiktoken==0.4.0
time-machine==2.13.0
toml==0.10.2
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
ultralytics==8.0.58
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.31.1
virtualenv==20.24.6
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.2
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.20.0